kegiatan posyandu

+++ application/views/monitoring/kegiatan_posyandu/index.php
@@ -12,9 +12,11 @@
                <a type="button" class="btn btn-success float-right ml-2 btn-block" href="<?= base_url('monitoring/kegiatan_posyandu/pdf') ?>">
                   <i class="fas fa-print"></i> PDF
                </a>
-               <button type="button" class="btn btn-primary float-right btn-block" data-toggle="modal" data-target="#addModal">
-                  <i class="fas fa-plus"></i> Tambah
-               </button>
+               <?php if ($role == 2) : ?>
+                  <button type="button" class="btn btn-primary float-right btn-block" data-toggle="modal" data-target="#addModal">
+                     <i class="fas fa-plus"></i> Tambah
+                  </button>
+               <?php endif; ?>
             </div>
          </div>
       </div>
@@ -31,6 +33,7 @@
                      <th>Sasaran</th>
                      <th>Parameter Keberhasilan</th>
                      <th>Photo</th>
+                     <th>Disetujui</th>
                      <th>Aksi</th>
                   </tr>
                </thead>
@@ -44,9 +47,15 @@
                         <td><?= $field['sasaran'] ?></td>
                         <td><?= $field['parameter_keberhasilan'] ?></td>
                         <td><img src="<?= base_url('assets/img/kegiatan_posyandu/') . $field['photo'] ?>" width="150"></td>
+                        <td><?= $field['is_verified'] ? 'Sudah' : 'Belum' ?></td>
                         <td>
-                           <button type="button" class="btn btn-primary btn-sm" data-toggle="modal" data-target="#editModal" onclick="getData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Edit</button>
-                           <button type="button" class="btn btn-danger btn-sm" data-toggle="modal" data-target="#deleteModal" onclick="deleteData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Hapus</button>
+                           <?php if ($role == 2) : ?>
+                              <button type="button" class="btn btn-primary btn-sm" data-toggle="modal" data-target="#editModal" onclick="getData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Edit</button>
+                              <button type="button" class="btn btn-danger btn-sm" data-toggle="modal" data-target="#deleteModal" onclick="deleteData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Hapus</button>
+                           <?php endif; ?>
+                           <?php if ($role == 8) : ?>
+                              <button type="button" class="btn btn-info btn-sm" data-toggle="modal" data-target="#prosesModal" onclick="deleteData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Proses</button>
+                           <?php endif; ?>
                         </td>
                      </tr>
                   <?php endforeach; ?>
